refactor(canvas): move plugin config to its owner
