+++ apps/canvas-workspace/harness/knowledge/plugin-market.md
@@ -36,7 +36,7 @@ The market is a Canvas application feature, not the engine's `EnginePlugin` or `
 | Install/trust state and managed storage paths | `src/main/plugin-market/store.ts` |
 | Standard v1 MCP to Pulse MCP config conversion | `src/main/plugin-market/mcp-adapter.ts` |
 | Normalized package to legacy Canvas registries/config adapter | `src/main/plugin-market/canvas-package-adapter.ts` |
-| Canvas plugin directory/config SSOT and skill sources | `src/main/settings/canvas-plugins-config.ts` |
+| Canvas plugin directory/config SSOT and skill sources | `src/main/plugin-market/config.ts` |
 | IPC registration and preload bridge | `src/main/plugin-market/ipc.ts`, `src/preload/bridge/plugin-market.ts`, `src/preload/index.ts` |
 | Route, state, filters, rows and dialogs | `src/renderer/src/modules/plugin-market/`, wired by `src/renderer/src/app/App/index.tsx` |
 | Installed-plugin `@` mentions and request-context collection | `src/renderer/src/modules/chat/mentions/pluginMentionItems.ts`, `modules/chat/components/ChatComposer/useChatComposerInput.ts` |
@@ -163,7 +163,7 @@ node scripts/harness/run-harness-check.mjs --path apps/canvas-workspace/src/main
 Focused iteration:
 
 ```bash
-pnpm --filter canvas-workspace exec vitest run src/main/plugin-market src/main/settings/__tests__/canvas-plugins-config.test.ts src/renderer/src/modules/plugin-market
+pnpm --filter canvas-workspace exec vitest run src/main/plugin-market src/renderer/src/modules/plugin-market
 pnpm --filter canvas-workspace typecheck
 node apps/canvas-workspace/harness/tools/describe-canvas.mjs
 ```
