+++ apps/canvas-workspace/harness/knowledge/security-posture.md
@@ -75,7 +75,7 @@ These make on-disk files an execution or injection surface:
   provider flow (`canvas-agent.ts:683-688`). Same SSRF/spawn shape as
   engine's posture doc describes.
 - **Canvas node plugins** (`canvas-plugins.json` + `pluginDirs`,
-  `src/main/settings/canvas-plugins-config.ts`) load external plugin code
+  `src/main/plugin-market/config.ts`) load external plugin code
   into main and renderer via manifests/registries.
 - **Model config is the positive example:** it stores only env-var *names*
   for API keys, never secret values (`src/main/agent/model/config.ts`).
