+++ apps/canvas-workspace/src/main/__tests__/file-size-governance.test.ts
@@ -26,7 +26,7 @@ const CURRENT_OVER_500_BASELINE: Record<string, number> = {
   'src/main/runtime/mcp-server.ts': 652,
   'src/main/agent/model/config.ts': 599,
   'src/plugins/main/dynamic-app/tools.ts': 593,
-  'src/main/settings/canvas-plugins-config.ts': 558,
+  'src/main/plugin-market/config.ts': 502,
   // 605→636 (2026-07-17, drift recorded): grew via master work (#806
   // session-restore fix) that never ran this suite (no automatic trigger).
   // Raised to measured; must-not-grow applies from 636.
